src/requirements.txt
aiohttp==3.13.2
beautifulsoup4==4.14.2
curl-cffi==0.13.0
paho-mqtt==2.1.0
requests==2.32.5
requests-toolbelt==1.0.0
sseclient==0.0.27
urllib3==2.5.0
websockets==15.0.1
zeroconf==0.148.0